update for FileFetcher's futures not being allowed to be shared across runtimes
