+++ cli/lsp/jsr.rs
@@ -17,7 +17,7 @@ use super::search::PackageSearchApi;
 
 #[derive(Debug)]
 pub struct CliJsrSearchApi {
-  file_fetcher: FileFetcher,
+  file_fetcher: Arc<FileFetcher>,
   resolver: JsrFetchResolver,
   search_cache: DashMap<String, Arc<Vec<String>>>,
   versions_cache: DashMap<String, Arc<Vec<Version>>>,
@@ -25,7 +25,7 @@ pub struct CliJsrSearchApi {
 }
 
 impl CliJsrSearchApi {
-  pub fn new(file_fetcher: FileFetcher) -> Self {
+  pub fn new(file_fetcher: Arc<FileFetcher>) -> Self {
     let resolver = JsrFetchResolver::new(file_fetcher.clone());
     Self {
       file_fetcher,
@@ -56,11 +56,15 @@ impl PackageSearchApi for CliJsrSearchApi {
     }
     let mut search_url = jsr_api_url().join("packages")?;
     search_url.query_pairs_mut().append_pair("query", query);
-    let file = self
-      .file_fetcher
-      .fetch(&search_url, &PermissionsContainer::allow_all())
-      .await?
-      .into_text_decoded()?;
+    let file_fetcher = self.file_fetcher.clone();
+    // spawn due to the lsp's `Send` requirement
+    let file = deno_core::unsync::spawn(async move {
+      file_fetcher
+        .fetch(&search_url, &PermissionsContainer::allow_all())
+        .await?
+        .into_text_decoded()
+    })
+    .await??;
     let names = Arc::new(parse_jsr_search_response(&file.source)?);
     self.search_cache.insert(query.to_string(), names.clone());
     Ok(names)
