+++ cli/npm/managed/mod.rs
@@ -154,7 +154,6 @@ fn create_inner(
   let fs_resolver = create_npm_fs_resolver(
     fs.clone(),
     npm_cache.clone(),
-    npm_rc.clone(),
     &text_only_progress_bar,
     resolution.clone(),
     tarball_cache.clone(),
@@ -179,7 +178,6 @@ fn create_inner(
     maybe_lockfile,
     npm_api,
     npm_cache,
-    npm_rc,
     package_json_deps_installer,
     resolution,
     tarball_cache,
@@ -268,7 +266,6 @@ pub struct ManagedCliNpmResolver {
   maybe_lockfile: Option<Arc<Mutex<Lockfile>>>,
   npm_api: Arc<CliNpmRegistryApi>,
   npm_cache: Arc<NpmCache>,
-  npm_rc: Arc<ResolvedNpmRc>,
   package_json_deps_installer: Arc<PackageJsonDepsInstaller>,
   resolution: Arc<NpmResolution>,
   tarball_cache: Arc<TarballCache>,
@@ -292,7 +289,6 @@ impl ManagedCliNpmResolver {
     maybe_lockfile: Option<Arc<Mutex<Lockfile>>>,
     npm_api: Arc<CliNpmRegistryApi>,
     npm_cache: Arc<NpmCache>,
-    npm_rc: Arc<ResolvedNpmRc>,
     package_json_deps_installer: Arc<PackageJsonDepsInstaller>,
     resolution: Arc<NpmResolution>,
     tarball_cache: Arc<TarballCache>,
@@ -305,7 +301,6 @@ impl ManagedCliNpmResolver {
       maybe_lockfile,
       npm_api,
       npm_cache,
-      npm_rc,
       package_json_deps_installer,
       text_only_progress_bar,
       resolution,
@@ -576,7 +571,6 @@ impl CliNpmResolver for ManagedCliNpmResolver {
       create_npm_fs_resolver(
         self.fs.clone(),
         self.npm_cache.clone(),
-        self.npm_rc.clone(),
         &self.text_only_progress_bar,
         npm_resolution.clone(),
         self.tarball_cache.clone(),
@@ -586,7 +580,6 @@ impl CliNpmResolver for ManagedCliNpmResolver {
       self.maybe_lockfile.clone(),
       self.npm_api.clone(),
       self.npm_cache.clone(),
-      self.npm_rc.clone(),
       self.package_json_deps_installer.clone(),
       npm_resolution,
       self.tarball_cache.clone(),
