+++ cli/npm/managed/resolvers/mod.rs
@@ -7,7 +7,6 @@ mod local;
 use std::path::PathBuf;
 use std::sync::Arc;
 
-use deno_npm::npm_rc::ResolvedNpmRc;
 use deno_npm::NpmSystemInfo;
 use deno_runtime::deno_fs::FileSystem;
 
@@ -25,7 +24,6 @@ use super::resolution::NpmResolution;
 pub fn create_npm_fs_resolver(
   fs: Arc<dyn FileSystem>,
   npm_cache: Arc<NpmCache>,
-  npm_rc: Arc<ResolvedNpmRc>,
   progress_bar: &ProgressBar,
   resolution: Arc<NpmResolution>,
   tarball_cache: Arc<TarballCache>,
